Add `pkg_test.go` files for every package
Unless a `pkg_test.go` file is present in a `pkg` package, go toolchain
will not consider that package for the purposes of coverage metrics. This
is a regression and will be fixed in go1.22 but until then we need these
files in each package to keep the coverage report honest.