Update codecov.yml to omit generated code
Everytime we add or modify a CRD, we will fail to meet generated
code coverage. The onus of ensuring the correctness of generated
code lies on the generator itself. We can, in future, think about
generating tests for this code (trust but verify) but, for the time
being, it's best to exclude this from metrics.