Check-in v1.3.0 manifests in test/e2e/data and bump versions (#367)
v1.3.0 manifests are to to spin up a point in time infrastructure from
a particular release. We also updated the versions to ensure there's a
version that e2e tests can target. We have also added a
'release-blocker' comment everywhere that needs to be modified to ensure
a release does not break e2es.