Update go directive in go.mod to 1.21
Based on https://endoflife.date/go go 1.21 is the lowest still supported version of go. Go EOL policy is basically supporting two recent minor versions. Currently, CAPX main branch and 1.4.x releases have go directive set to 1.22 which is fine as long as we build it but is a problem when someone tries to build our repo with a supported but not latest version of go (e.g. Openshift building CAPX with their go 1.21 toolchain). We should change our go version to 1.21 and keep it set to the lowest supported version of go. i.e. we should stick to 1.21 until 1.23 is released and then move to 1.22 once 1.23 gets released and so on.