Remove stale E2E test
With https://github.com/nutanix-cloud-native/cluster-api-provider-nutanix/pull/400
we made prismCentral a mandatory field in NutanixCluster. In the PR we did not
remove the E2E test that was covering the older behavior of successfully
creating a NutanixCluster without credentials using the default credentials
from the CAPX controller. As a result, the test is now failing on both periodic
jobs since and presubmits on new PRs with the following error:

```
NutanixCluster.infrastructure.cluster.x-k8s.io \"cluster-ntnx-client-qwqlsf\" is invalid: spec.prismCentral: Required value
```

This PR removes the test that tests the stale example that is no longer supported.