fix(bug): Remove crude check for prism central version
Earlier we had a crude check on prism central version which checked
if prism central version is higher than 2024.1. We remove this check
and assume that the underlying prism central is within support lifespan.
This fixes the false positive issues when this check failed for newer
PC versioning scheme of following AOS version i.e. PC7.3.